refactor: 미확인 알림 조회 API count 쿼리로 변경 (#250)

+++ module-domain/src/main/java/com/depromeet/notification/port/in/usecase/GetFollowLogUseCase.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface GetFollowLogUseCase {
     List<FollowLog> getFollowLogs(Long memberId, LocalDateTime cursorCreatedAt);
 
-    int getUnreadFollowLogCount(Long memberId);
+    Long getUnreadFollowLogCount(Long memberId);
 }
