+++ module-domain/src/main/java/com/depromeet/notification/port/in/usecase/GetReactionLogUseCase.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface GetReactionLogUseCase {
     List<ReactionLog> getReactionsLogs(Long memberId, LocalDateTime cursorCreatedAt);
 
-    int getUnreadReactionLogCount(Long memberId);
+    Long getUnreadReactionLogCount(Long memberId);
 }
